# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Planventory Inventory API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Planventory-Proprietary
  description: |
    Inventory management API. All monetary amounts are integers in the smallest currency unit.
    Successful and error responses use a common envelope with `success`, optional `data`, optional `error`, and optional `meta`.
servers:
  - url: /
    description: API base (prepend your host)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate access token using httpOnly refresh cookie
      operationId: refreshToken
      security: []
      description: Uses the refresh token sent in an httpOnly cookie; no JSON body.
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/auth/change-password:
    patch:
      tags: [Auth]
      summary: Change password for the authenticated user
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptySuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/purchase-orders/{id}/receive:
    post:
      tags: [PurchaseOrders]
      summary: Record received quantities for a purchase order
      operationId: receivePurchaseOrder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PurchaseOrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivePurchaseOrderRequest"
      responses:
        "200":
          description: Receipt processed; inventory updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderSingleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Purchase order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
...
    PurchaseOrderId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Validation issues or structured context
          oneOf:
            - type: array
              items: {}
            - type: object
              additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
    EmptySuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
...
    RefreshResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required: [accessToken]
          properties:
            accessToken:
              type: string
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
...
    PurchaseOrderStatus:
      type: string
      enum: [DRAFT, SENT, PARTIALLY_RECEIVED, RECEIVED, CANCELLED]
    PurchaseOrderLine:
      type: object
      required: [variantId, orderedQty, unitCost]
      properties:
        variantId:
          type: string
        orderedQty:
          type: integer
          minimum: 1
        unitCost:
          type: integer
          description: Cost per unit in smallest currency unit
    PurchaseOrder:
      type: object
      required: [id, supplierId, status, items, notes]
      properties:
        id:
          type: string
        supplierId:
          type: string
        status:
          $ref: "#/components/schemas/PurchaseOrderStatus"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrderLine"
        notes:
          type:
            - string
            - "null"
...
    PurchaseOrderSingleResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/PurchaseOrder"
...
    ReceivePurchaseOrderItem:
      type: object
      required: [variantId, receivedQty]
      properties:
        variantId:
          type: string
        receivedQty:
          type: integer
          minimum: 0
    ReceivePurchaseOrderRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ReceivePurchaseOrderItem"
...